Correct in_flight tracking + Monitor Spiders independently in the loop
Correct in_flight request decrease tracking place (before if check in the queue, when error happened with request, will not decrease in the right place).
Monitor Spiders independently in the loop: Before this logic located on the item receive logic okay, which on monitor when receive an item from the queue and also the item is not error.